feat:別ファイルに分割したのはいいんだけど、そもそもこういう使い方じゃない気がする...
